moved the ShoppingListForm into own component to work on mui with
